fixed bug in compoundf
The threshold for 2^t rounding to nearest was too large.
But since we compared the encoding of 2*t.u, we should have
multiplied the constant by 2. Since this was not done, the
comparison was done with a much smaller threshold (thus there
was no real bug, but the comment was wrong and the code was inefficient).